RF: Move server parts to their own script
This allows us to keep the module-loading parts of pelita-player
self-contained. Ideally, we only want to interact with pelita-player via
subprocesses to avoid contaminating our own namespace with team module
imports.
