+++ pyproject.toml
@@ -43,6 +43,7 @@ pelita = "pelita.scripts.pelita_main:main"
 pelita-tournament = "pelita.scripts.pelita_tournament:main"
 pelita-tkviewer = "pelita.scripts.pelita_tkviewer:main"
 pelita-player = "pelita.scripts.pelita_player:main"
+pelita-server = "pelita.scripts.pelita_server:main"
 pelita-createlayout = "pelita.scripts.pelita_createlayout:main"
 
 [tool.aliases]
